Module rewards::pallet

The module that hosts all the FRAME types needed to add this pallet to a runtime.

Re-exports

pub use __tt_error_token_1 as tt_error_token;

pub use __tt_default_parts_7 as tt_default_parts;

Structs

GenesisConfig

Can be used to configure the genesis state of this pallet.

Pallet

The pallet implementing the on-chain logic.

Enums

Call

Contains one variant per dispatchable that can be called by an extrinsic.

Error

Custom dispatch errors of this pallet.

Event

The event emitted by this pallet.

Traits

Config

Configuration trait of this pallet.

Type Definitions

CurrentClaimableVal

Amount of VAL currently being vested (aggregated over the previous period of 14,400 blocks)

EthAddresses

All addresses are split in batches, AddressBatches maps batch number to a set of addresses

MigrationPending

A flag indicating whether VAL rewards data migration has been finalized

ModuleDeprecated

Type alias to Pallet, to be used by construct_runtime.

PswapFarmOwners

PswapWaifuOwners

ReservesAcc

TotalClaimableVal

Total amount of VAL that can be claimed by users at current point in time

TotalValRewards

Total amount of VAL rewards either claimable now or some time in the future

UmiNftClaimed

Stores whether address already claimed UMI NFT rewards.

UmiNftReceivers

UMI NFT receivers storage

UmiNfts

The storage of available UMI NFTs.

ValBurnedSinceLastVesting

Amount of VAL burned since last vesting

ValOwners

A map EthAddresses -> RewardInfo, that is claimable and remaining vested amounts per address